Initial commit - exploring using a GNN to predict PCBA AID 577 "Inhibitors of West Nile Virus NS2bNS3 Proteinase" results from molecular structure
Trying to apply a similar network to https://github.com/willy-b/tiny-GIN-for-ogbg-molhiv/ to a West Nile Virus related dataset.

I will start by considering the West Nile Virus related dataset "HTS to identify Inhibitors of West Nile Virus NS2bNS3 Proteinase" and predicting such NS2bNS3 protease inhibitors as they were considered likely at the time of that datasets publication to be WNV antiviral drug candidates:
![PubChem BioAssay AID 577 - HTS to identify Inhibitors of West Nile Virus NS2bNS3 Proteinase](pcba_aid_577_description_snippet.png)

We can put that into OGB format for consistency and test a similar tiny Graph Isomorphism Network (GIN) on it.

NOTE THERE IS NO COMPETITION ASSOCIATED WITH THIS DATASET THAT I AM AWARE OF (this is NOT the ogbg-molpcba task which is 128 different PCBA assays, none of which are 577), THIS IS JUST FOR INTEREST IN THE PROBLEM UNLIKE THE OGBG-MOLHIV REPO