Improve dataset prep routine for reading the AID_577_datatable.csv (not varying, fixed in repo, tested with code)
We are ONLY reading AID_577_datatable.csv which is committed to our repo by the prep.py author so not likely to change in any surprising way and break anything.
Still, even in the draft branch, we should not do sloppy processing so just skipping the 3 empty column lines after the header instead of putting scary looking continues into the data processing loop which look as if we might be silently skipping data or misaligning the columns (it is not, but if applied to some other data as it was drafted it could).

Has no change on anything, not even splitting with fixed random seed as it did not change which lines were read or how they are handled, just safer for corrupted data in the future.