add `--active_class_weight` argument for allowing reweighting to account for class imbalance
Optionally, allow reweighting of loss to account for class imbalance.
Default of 1.0 has no effect,
setting to 1/active_molecule_prevalence would encourage recall of active molecules same as inactive at cost of precision.
When judging by ROCAUC and/or AP using the raw y prediction scores
to rank the results, this is not necessary but may of interest
especially if using output as a classifier on a single example
or online rather than ranking a list.

I am not using this yet but figured may be useful for comparison / exploration.