Update main_gin.py - remove outdated redundant comment as eval_on_test_set argument was added in https://github.com/willy-b/tiny-GIN-for-WNV/pull/15/commits/6bcf3cc3674b3af4cd0deb6bcd6af119700d77ed
Back in https://github.com/willy-b/tiny-GIN-for-WNV/pull/15/commits/6bcf3cc3674b3af4cd0deb6bcd6af119700d77ed , the eval_on_test_set argument was added allowing test set evaluation. 

There was a comment from early days on a different line mentioning that an argument called "hide_test_metric" might be added once test set evaluation was supported, but that is not the approach that was taken (instead default to NOT evaluating on the test set and showing the result), so remove that outdated comment (not shown to user, just in the code).